Allow not hiding ticks in Greek type variables (#3)
Summary:
This patch allows concealing `'a` as `'α` instead of just `α`. This
prevents the line width from varying, which is nice.

It's opt-in and does not change the default behavior; to use it, add
```
let g:sml_greek_tyvar_show_tick = 1
```
to your vimrc.

Test Plan:
Write `val x : 'a list = nil` in an SML file, and watch it be concealed
properly.

Write `val x : 'xyz list = nil` and watch it be highlighted but not concealed.