Consider `local` in indentation (#4)
### Summary

This commit updates the indentation settings to properly handle `local`
blocks, which were previously not considered at all. I basically added
`local` everywhere I saw `let`, and then checked that the semantics made
sense! :-)

### Test Plan

Entering the code

```sml
structure S = struct
  structure T = struct
    local
      type number = int
      fun f (x : number : number) =
        let
          val y : number = x + 1
        in
          1 * y * 1
        end
    in
      val g : int -> int = f
    end
  end
end
```

now causes it to be indented as above, whereas before it looked like

```sml
structure S = struct
  structure T = struct
    local
    type number = int
    fun f (x : number) : number =
      let
        val y : number = x + 1
      in
        1 * y * 1
      end
      in
    val g : int -> int = f
  end
end
end
```

(This was happening because the `end` was being matched to the previous
structure definition instead of the previous `local` block.)

Note that you have to enter this code manually, as opposed to using `=`
to indent it after the fact; re-indenting was already pretty broken (try
it yourself: `=ip`), and this patch does not fix that.