def-use: Put def's before uses
This way, if we grep the file from the top for a location to find it's
def, and the location we're looking for happens to BE the def, it's the
first result we find.

Otherwise, what could happen would be that we lookup the def and get
back the use! There's no type information at the use, so we'd be hosed.
Thus, it has to come before all the uses.