Don't quit :SMLReplStart if first build fails
Previously, we were using `sml foo.sml` or `sml -m foo.cm` to launch the
REPL. This method has the unfortunate side effect that if there are
errors when opening the REPL, the REPL will quit until you fix the
errors.

Now, we always launch the REPL without command line args, and issue a
bettersml#repl#bReplBuild() after the REPL has started, which delegates
to a normal `use` / `CM.make` call for the first build. This way it
doesn't quit the buffer.